making new changes to make the new module implementation more stable

+++ Interfaces/IModule.cs
@@ -3,8 +3,8 @@ namespace Voyage;
 public interface IModule
 {
       public void Refresh();
-      public void ToggleElement(ushort index, bool toggle);
-      public void ToggleElementSelection(ushort[] group, bool toggle);
+      public void ToggleElement(int index, bool toggle);
+      public void ToggleElementSelection(int[] group, bool toggle);
 }
 
 public interface IModule<T> : IModule
